requirements.txt
flask==2.0.3
gunicorn==19.9.0
PyYAML==5.1
pytest==4.5.0
pytest-flask==0.11.0
pytest-mock==1.10.0